入れ替えに備えてnask_defs.hppをdeprecatedとする
